requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
praw==7.1.2
prawcore==1.5.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.3
websocket-client==0.57.0
whitenoise==5.2.0
